Bug fix
- fixed some bugs

+++ src/commands/removePublicRole.ts
@@ -12,7 +12,7 @@ const commandData: commandI = {
         const roletoRemove = interaction.options.getRole("roletoremove");
         const prisma = new PrismaClient();
         await interaction.reply({
-            content: `${roletoRemove} was added to public role list.`,
+            content: `${roletoRemove} was removed from public role list.`,
             ephemeral: true,
         });
 
@@ -37,12 +37,21 @@ const commandData: commandI = {
             return;
         }
 
-        await prisma.roleToChoose.deleteMany({
+        const role = await prisma.roleToChoose.findFirst({
             where: {
                 serverId: Number(DBserverData.id),
                 name: roletoRemove.name,
                 roleId: roletoRemove.id,
             },
+        });
+
+        if (role === null) {
+            interaction.editReply("Role is not in the list.");
+            return;
+        }
+
+        await prisma.roleToChoose.deleteMany({
+            where: role,
         });
     },
     initOptions: (slashCommandBuild: SlashCommandBuilder) => {
@@ -51,6 +60,7 @@ const commandData: commandI = {
             option.setDescription(
                 "Role that be added to the public roles list.",
             );
+            option.setRequired(true);
 
             return option;
         });
